+++ src/components/ChatInput/ChatInput.js
@@ -1,7 +1,6 @@
 import { Box, Button, Icon, ActionButton } from '@rocket.chat/fuselage';
 import React, { useState, useContext, useRef, useEffect } from 'react';
 import { useToastBarDispatch } from '@rocket.chat/fuselage-toastbar';
-import emojione from 'emoji-toolkit';
 import styles from './ChatInput.module.css';
 import RCContext from '../../context/RCInstance';
 import {
@@ -16,6 +15,7 @@ import MembersList from '../Mentions/MembersList';
 import mentionmemberStore from '../../store/mentionmemberStore';
 import { searchToMentionUser } from '../../lib/searchToMentionUser';
 import TypingUsers from '../TypingUsers';
+import { parseEmoji } from '../../lib/emoji';
 
 const ChatInput = () => {
   const { RCInstance } = useContext(RCContext);
@@ -162,17 +162,6 @@ const ChatInput = () => {
     } catch (e) {
       console.error(e);
     }
-  };
-
-  const parseEmoji = (text) => {
-    const regx = /:([^:]*):/g;
-    const regx_data = text.match(regx);
-    if (regx_data) {
-      const result = regx_data[regx_data.length - 1];
-      const d = emojione.shortnameToUnicode(result);
-      if (d !== undefined) text = text.replace(result, d);
-    }
-    return text;
   };
 
   return (
